Refactor extcallset interpreter tests
