+++ crates/lib/vm-interpreter-extcallset/Cargo.toml
@@ -15,7 +15,8 @@ derive-where = { workspace = true }
 thiserror = { workspace = true }
 
 [dev-dependencies]
+waymark-vm-bytecode = { workspace = true }
 waymark-vm-interpreter-coreset = { workspace = true }
 waymark-vm-runtime = { workspace = true }
-waymark-vm-runtime-promise-value = { workspace = true }
 waymark-vm-runtime-test = { workspace = true }
+waymark-vm-runtime-value = { workspace = true }
